net: move CBanDB and CAddrDB out of net.h/cpp
