+++ src/makefile.bsd
@@ -103,6 +103,7 @@ OBJS= \
     obj/version.o \
     obj/checkpoints.o \
     obj/netbase.o \
+    obj/addrdb.o \
     obj/addrman.o \
     obj/crypter.o \
     obj/key.o \
