+++ src/makefile.mingw
@@ -72,6 +72,7 @@ OBJS= \
     obj/version.o \
     obj/checkpoints.o \
     obj/netbase.o \
+    obj/addrdb.o \
     obj/addrman.o \
     obj/crypter.o \
     obj/key.o \
