+++ src/makefile.osx
@@ -76,6 +76,7 @@ OBJS= \
     obj/support/cleanse.o \
     obj/checkpoints.o \
     obj/netbase.o \
+    obj/addrdb.o \
     obj/addrman.o \
     obj/base58.o \
     obj/crypter.o \
